+++ app/icm20948/icm20948_dmp.hpp
@@ -25,7 +25,7 @@ namespace ICM20948 {
 
     constexpr auto ACCEL_SCALE = 8.192F / 1000.0F * 9.81F;
     constexpr auto GYRO_SCALE = 16.4F;
-    constexpr auto QUAT_SCALE = static_cast<std::float32_t>(1UL << 30UL);
+    constexpr auto QUAT_SCALE = static_cast<std::float64_t>(1UL << 30UL);
 
     struct ICM20948_DMP {
     public:
@@ -40,12 +40,12 @@ namespace ICM20948 {
 
         ~ICM20948_DMP() noexcept;
 
-        std::optional<Vec3D<std::float32_t>> get_roll_pitch_yaw() noexcept;
-        std::optional<std::float32_t> get_roll() noexcept;
-        std::optional<std::float32_t> get_pitch() noexcept;
-        std::optional<std::float32_t> get_yaw() noexcept;
+        std::optional<Vec3D<std::float64_t>> get_roll_pitch_yaw() noexcept;
+        std::optional<std::float64_t> get_roll() noexcept;
+        std::optional<std::float64_t> get_pitch() noexcept;
+        std::optional<std::float64_t> get_yaw() noexcept;
 
-        std::optional<Quat3D<std::float32_t>> get_quaternion_scaled() noexcept;
+        std::optional<Quat3D<std::float64_t>> get_quaternion_scaled() noexcept;
 
     private:
         std::optional<Quat3D<std::int32_t>> get_quaternion_raw() noexcept;
